revised according to PR #10 comments.
