+++ api/blueprints/upload/controllers.py
@@ -4,16 +4,12 @@ from api.credentials import host_path
 from flask import Blueprint, request, make_response
 from http import HTTPStatus
 from api import require_apikey
-import hashlib
+from api.apiutils import hash_file, valid_file_type
 
 from werkzeug.utils import secure_filename
 
 upload = Blueprint('upload', __name__)
 
-# Our buffer size for file reads is
-BUF_SIZE = 2 << ((10 * 1) + 4)  # 16 KB
-MAX_SIZE = 2 << ((10 * 2) + 1)  # 2 MB
-ALLOWED_EXTENSIONS = {'gif', 'png', '.jpg'}
 IMAGES_BASE = "https://www.culturemesh.com/user_images/"
 
 @upload.route("/ping")
@@ -57,33 +53,3 @@ def upload_image():
     # Return new url.
     return IMAGES_BASE + directory_name + "/" + file_name
 
-
-def hash_file(file):
-    """
-    Generates a string hex hash (using md5) of the image file. We use a buffer to separate the file into
-    memory-manageable chunks.
-    This also throws TooLargeImageException if the file buffer manages to read more than 2MB of data.
-    :param file: should be a python file.
-    :return: string of hash in hex.
-    """
-    md5 = hashlib.md5()
-    data = file.read(BUF_SIZE)
-    file_size = BUF_SIZE
-    while data:
-        md5.update(data)
-        data = file.read(BUF_SIZE)
-        file_size += BUF_SIZE
-        if file_size >= MAX_SIZE:
-            raise MemoryError("file size too large")
-    # Reset cursor for file write
-    file.seek(0, 0)
-    return md5.hexdigest()
-
-
-def valid_file_type(file):
-    """
-    Checks if file type is either PNG, JPG, or GIF, which are our valid image formats.
-    :param file: python file.
-    :return: true if file is .png, .jpg, or .gif, false otherwise.
-    """
-    return file.filename.split(".")[-1] in ALLOWED_EXTENSIONS
